requirements.txt
langchain==0.2.13
langgraph==0.2.0
chromadb==0.5.4
sentence-transformers==3.0.1
google-generativeai==0.7.2
pymupdf==1.24.9
beautifulsoup4==4.12.3
langchain-community==0.2.12
fastapi==0.114.0
uvicorn[standard]==0.30.0
streamlit==1.37.0
mlflow==2.15.1
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4